Move the old Production to the GravitationalWaveTransient class.
